Move handleRecordOn to @utils/data/form/updates.ts
